requirements.txt
accelerate==1.11.0
aiofiles==25.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bitsandbytes==0.48.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.120.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
h11==0.16.0
httptools==0.7.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.2
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
sympy==1.14.0
tifffile==2025.10.16
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.222
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
supervision
ffmpeg-python
accelerate==1.11.0
aiofiles==25.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
bitsandbytes==0.48.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
defusedxml==0.7.1
easyocr==1.7.2
et_xmlfile==2.0.0
fastapi==0.120.1
ffmpeg-python==0.2.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
future==1.0.0
h11==0.16.0
httptools==0.7.1
huggingface-hub==0.36.0
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.2
pyclipper==1.3.0.post6
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.1
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.3
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
supervision==0.26.1
sympy==1.14.0
tifffile==2025.10.16
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.222
ultralytics-thop==2.0.18
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
